Kah-Auto Storage
Made Registration activity with login and sign-up fragment.
Made a Cars service for  foregroundService.
Made User - data class, usersDao, usersDatabase.

--------
Still need to solve a problem on Registration activity
on fun isUserLegit(), Got a crash when trying to read from Repository.
^
Same as this problem but on the Notification manager in
DisplayOver24h. always getting null from car list in the repository,